Changed from natdin to biblatex.
Package natdin is exchanged with biblatex. Therefore the package
csquotes has to be included. Now you have to use biber instead of
bibtex for compiling. The commands like "\Zitat" are now using
\parencite{} instead of \citep*{}. Two filetypes are added to the
gitignore which are generated during compilation because of biblatex
and biber. Old packages and lines in the main file are commented out.
The old commands are deleted.